Commit direct from IDEA

+++ src/learnJavaAtPrison/ImplementationTricks.java
@@ -4,11 +4,11 @@ public class ImplementationTricks {
     public static void main(String[] args) {
         int primer=1;
         System.out.println("begining number is "+primer);
-        System.out.println("++i "+ (int)++primer);
-        System.out.println("i++ "+ (int)primer++);
-        System.out.println("i "+ (int)primer);
-        System.out.println("--i "+ (int)--primer);
-        System.out.println("i-- "+ (int)primer--);
+        System.out.println("++i "+ ++primer);
+        System.out.println("i++ "+ primer++);
+        System.out.println("i "+ primer);
+        System.out.println("--i "+ --primer);
+        System.out.println("i-- "+ primer--);
         System.out.println("i "+ primer);
     }
 }
